Feat: Add publication type selection screen
- Adds a preliminary step to the publication creation process to select the type of publication.
- Refactors the existing form to handle the pre-selected type.
- The user is now first presented with a choice of publication types (Motor, Regulador, Otro Repuesto) before entering the main publication form.
